Fixed GDT entry limit size

+++ kernel/include/kernel/gdt.h
@@ -5,7 +5,7 @@
 
 typedef struct GDT {
     uint32_t base;
-    uint32_t limit;
+    uint16_t limit;
     uint8_t access_byte;
     uint8_t flags;
 } GDT;
